+++ Łączenie/GeneticAlgorithm.hpp
@@ -11,7 +11,6 @@
 #include<sstream>
 #include<iostream>
 #include "generateRandom.hpp"
-//const unsigned int thread_number = 10;
 
 class GeneticAlgorithm
     {
@@ -61,10 +60,10 @@ class GeneticAlgorithm
         Chromosome* restart(Chromosome* start_object ,double min_diffrence_between_generations_best_,unsigned int NUMBER_OF_REPETED_SUBOPTIMAL_SOLUTION_ = 1, unsigned int max_of_iteration_ = 20,unsigned int population_size_ = 20,unsigned int max_best_ = 3, std::vector<Chromosome*> population_ = std::vector<Chromosome*>());
 
         //setters
-        void setMaxOfIteration(int max_of_iteration_);
-        void setMinDifferenceBetweenGenarationsBest(int min_diffrence_between_generations_best_);
-        void setMaxPopulationSize(int max_population_size_);
-        void setMaxBest(int max_best_);
+        void setMaxOfIteration(unsigned int max_of_iteration_);
+        void setMinDifferenceBetweenGenarationsBest(double min_diffrence_between_generations_best_);
+        void setMaxPopulationSize(unsigned int max_population_size_);
+        void setMaxBest(unsigned int max_best_);
 
         //out
         friend std::ostream& operator<<(std::ostream& out, GeneticAlgorithm& obj);
